bi-copilot/backend/requirements.txt
fastapi
beautifulsoup4
google-genai
pandas
psycopg2-binary
pytest
requests
sqlalchemy
uvicorn
